+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # CHANGELOG
 
+## 1.0.0 - 2021-4-5
+
+### Fixes
+
+- Fix manifest license link (Issue #112)
+
+- Fix help document md to xml file compilation (Issue #90, #106)
+
+- Remove unnecessary assert in registry write code (Issue #119)
+
+### Changes
+
+- Add constructor for `SecretInformation` class that accepts metadata as a hash table (Issue #108)
+
+### New Features
+
 ## 0.9.1 - 2021-3-1
 
 ### Fixes